# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "Open Api",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/media": {
            "get": {
                "tags": [
                    "Media"
                ],
                "summary": "Get list of media files",
                "operationId": "b7198dfff17ffa70df56cd807f41abaa",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/query_limit"
                    },
                    {
                        "$ref": "#/components/parameters/query_page"
                    },
                    {
                        "$ref": "#/components/parameters/query_keyword"
                    },
                    {
                        "name": "folder_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "image",
                                "video",
                                "audio",
                                "document",
                                "file"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/MediaResource"
                                            }
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginationMeta"
                                        },
                                        "links": {
                                            "$ref": "#/components/schemas/PaginationLinks"
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "apiKey": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Media"
                ],
                "summary": "Upload a new media file",
                "operationId": "384996999b5dc974c3fd192c9328d087",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/MediaRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/MediaResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "apiKey": []
                    }
                ]
            }
        },
        "/api/v1/media/{id}": {
            "get": {
                "tags": [
                    "Media"
                ],
                "summary": "Get media details by id",
                "operationId": "3573334ce3674dd235f8a91c9855bf09",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/MediaResource"
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Media not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "apiKey": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Media"
                ],
                "summary": "Update media info (name, etc)",
                "operationId": "e500083a6b6069636ed730c0ce31a524",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "folder_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/MediaResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Media not found"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "apiKey": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Media"
                ],
                "summary": "Delete a media file",
                "operationId": "405a4579bc2822b04a2d6df1487f1d12",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Media not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "apiKey": []
                    }
                ]
            }
        },
        "/api/v1/media/bulk": {
            "post": {
                "tags": [
                    "Media"
                ],
                "summary": "Bulk action on media files",
                "operationId": "e07b45c08c95a068265865dc7377ed08",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    },
                                    "action": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Bulk action successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "apiKey": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "MediaRequest": {
                "required": [
                    "file"
                ],
                "properties": {
                    "file": {
                        "description": "The file to upload",
                        "type": "string",
                        "format": "binary"
                    },
                    "folder_id": {
                        "description": "The ID of the folder where the media belongs (optional)",
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
...
            "MediaResource": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "readable_size": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string"
                    },
                    "mime_type": {
                        "type": "string"
                    },
                    "size": {
                        "type": "integer"
                    },
                    "path": {
                        "type": "string"
                    },
                    "url": {
                        "type": "string"
                    },
                    "extension": {
                        "type": "string"
                    },
                    "disk": {
                        "type": "string"
                    },
                    "is_directory": {
                        "type": "boolean"
                    },
                    "is_image": {
                        "type": "boolean"
                    },
                    "is_video": {
                        "type": "boolean"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "PaginationMeta": {
                "properties": {
                    "total_pages": {
                        "type": "integer"
                    },
                    "limit": {
                        "type": "integer"
                    },
                    "total": {
                        "type": "integer"
                    },
                    "page": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "PaginationLinks": {
                "properties": {
                    "self": {
                        "type": "string"
                    },
                    "first": {
                        "type": "string"
                    },
                    "prev": {
                        "type": "string"
                    },
                    "next": {
                        "type": "string"
                    },
                    "last": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
...
        "parameters": {
...
            "query_limit": {
                "name": "limit",
                "in": "query",
                "schema": {
                    "type": "integer"
                }
            },
            "query_page": {
                "name": "page",
                "in": "query",
                "schema": {
                    "type": "integer"
                }
            },
            "query_keyword": {
                "name": "keyword",
                "in": "query",
                "schema": {
                    "type": "string"
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "in": "header",
                "name": "Authorization"
            },
            "apiKey": {
                "type": "apiKey",
                "description": "A short description for security scheme",
                "name": "x-api-key",
                "in": "header"
            }
        }
    },
...
}